+++ src/aspire/utils/filters.py
@@ -1,13 +1,33 @@
 import inspect
+import math
 
 import numpy as np
 from scipy.interpolate import RegularGridInterpolator
 
 from aspire.utils import ensure
 from aspire.utils.coor_trans import grid_2d
-from aspire.utils.em import voltage_to_wavelength
 from aspire.utils.filter_to_fb_mat import filter_to_fb_mat
 from aspire.utils.matlab_compat import m_reshape
+
+
+def voltage_to_wavelength(voltage):
+    """
+    Convert from electron voltage to wavelength.
+    :param voltage: float, The electron voltage in kV.
+    :return: float, The electron wavelength in nm.
+    """
+    return 12.2643247 / math.sqrt(voltage * 1e3 + 0.978466 * voltage ** 2)
+
+
+def wavelength_to_voltage(wavelength):
+    """
+    Convert from electron voltage to wavelength.
+    :param wavelength: float, The electron wavelength in nm.
+    :return: float, The electron voltage in kV.
+    """
+    return (
+        -1e3 + math.sqrt(1e6 + 4 * 12.2643247 ** 2 * 0.978466 / wavelength ** 2)
+    ) / (2 * 0.978466)
 
 
 class Filter:
